# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Welcome to family budget API.",
        "title": "Family Budget API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Isfandiyor Zabirov",
            "url": "http://www.swagger.io/support",
            "email": "isfandiyor.zabirov.sh@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "10.192.9.236:8000",
    "basePath": "/api/v1",
    "paths": {
...
        "/login": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Регистрация и Авторизация"
                ],
                "summary": "Авторизация пользователя",
                "operationId": "authorize-user",
                "parameters": [
                    {
                        "description": "Даные для авторизации",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/middleware.Login"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/middleware.ResponseStruct"
                        }
                    },
                    "400": {
                        "description": "reason",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "reason",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "402": {
                        "description": "reason",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "middleware.Login": {
            "type": "object",
            "required": [
                "login",
                "password"
            ],
            "properties": {
                "captcha": {
                    "type": "string"
                },
                "device_uuid": {
                    "type": "string"
                },
                "login": {
                    "type": "string"
                },
                "otp": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "middleware.ResponseStruct": {
            "type": "object",
            "properties": {
                "access_expire": {
                    "type": "string"
                },
                "access_token": {
                    "type": "string"
                },
                "refresh_expire": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}